Follow the same file structure in ReadOnlyValueCollection
